Add Shelly3Pro
Add the Shelly3Pro, as found in the Shelly Pro 3. This is a custom
designed ESP32 board with 8MiB flash, and SMSC 8720A Ethernet PHY. All
other I/O is on separate boards via pin-headers and thus not relevant.

The board has no flashing circuity and requires 'manual' toggling of
reset and GPIO0.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>